+++ tests/test_flatten.py
@@ -51,7 +51,7 @@ def test_dontflatten_str():
 
 def test_flatten_str():
     nested = ["ab",3,("cd","e")]
-    assert list(flatten(nested, dontflatten=None)) == ["a","b",3,"c","d","e"]
+    assert list(flatten(nested, preserve=None)) == ["a","b",3,"c","d","e"]
 
 def test_dontflatten_bytes():
     nested = ["ab",3,(b"cd",b"e")]
@@ -59,4 +59,4 @@ def test_dontflatten_bytes():
 
 def test_flatten_bytes():
     nested = ["ab",3,(b"cd",b"e")]
-    assert list(flatten(nested, dontflatten=None)) == ["a","b",3,b"cd"[0],b"cd"[1],b"e"[0]]
+    assert list(flatten(nested, preserve=None)) == ["a","b",3,b"cd"[0],b"cd"[1],b"e"[0]]