+++ recurtools/__init__.py
@@ -1,6 +1,5 @@
 from .nested import nested
 from .utils import (
     flatten,
-    starchain,
     sumrecursive,
 )