+++ recurtools/__init__.py
@@ -1,5 +1,2 @@
 from .nested import nested
-from .utils import (
-    flatten,
-    sumrecursive,
-)
+from .utils import flatten