+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog: recurtools
 
 - rename chainanything to starchain
+- standardised on keywork argument `preserve`
 
 ## v0.4.0 - update to flatten (BREAKING)
 - **new keyword-only argument** `dontflatten` (optional): type which will not be flattened. Default: `(str, bytes)`.