+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog: recurtools
 
+## v1.0.0 - simplified interface around nested & flatten
+- documentation published to https://musicalninjadad.github.io/recurtools/
 - removed `sumrecursive()`see [Issue 11](https://github.com/MusicalNinjaDad/recurtools/issues/11)
 - removed `chainanything()` - some of the functionality may later be incorporated into `in nested.something()`
 - removed `indexrecursive()` function in preference for `nested.index()`.
 - removed `countrecursive()` function in preference for `nested.count()`.
 - removed `inrecursive()` function in preference for `in nested()`.
 - removed `lenrecursive()` function in preference for `len(nested())`. `countcontainers` not implemented in new version.
-- standardised on keywork argument `preserve`
+- standardised on keywork argument `preserve` for flatten. Will later be integrated into nested
 
 ## v0.4.0 - update to flatten (BREAKING)
 - **new keyword-only argument** `dontflatten` (optional): type which will not be flattened. Default: `(str, bytes)`.